# NotificationService/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Notification Service",
    "description": "A microservice for sending notifications to users within the FOOD TRACKER project.",
    "version": "1.0.0"
  },
  "host": "<domain pending>",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/user/{userId}": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get all notifications for a user",
        "description": "Get all notifications for a user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID"
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications found",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "example": "67844df7abdcdbf9d5c80487"
                  },
                  "userId": {
                    "type": "string",
                    "example": "15125645645645"
                  },
                  "userEmail": {
                    "type": "string",
                    "example": "user.email@email.com"
                  },
                  "content": {
                    "type": "string",
                    "example": "Example notification content, lorem ipsum 3 etc..."
                  },
                  "status": {
                    "type": "string",
                    "example": "unread"
                  },
                  "timestamp": {
                    "type": "string",
                    "example": "2025-01-02T21:15:03.839Z"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No notifications found for user or user does not exist"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "Delete all notifications for a user",
        "description": "Delete all notifications for a user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "All notifications for user deleted"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/send": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Create and send a notification",
        "description": "Create and send a notification",
        "parameters": [
          {
            "name": "userId",
            "in": "body",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5f9f1b3b3b6f6b0017f3b3b6"
            }
          },
          {
            "name": "userEmail",
            "in": "body",
            "description": "User email",
            "required": true,
            "schema": {
              "type": "string",
              "example": "user.email@email.com"
            }
          },
          {
            "name": "content",
            "in": "body",
            "description": "Notification content",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Notification email content, lorem ipsum etc..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification sent successfully"
          },
          "400": {
            "description": "Required: userId, content, userEmail"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/send/{notificationId}": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Send an existing notification",
        "description": "Send an existing notification",
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Notification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Notification sent successfully"
          },
          "404": {
            "description": "Notification not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/read/{notificationId}": {
      "put": {
        "tags": [
          "Notifications"
        ],
        "summary": "Update notification to read",
        "description": "Update notification to read",
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Notification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Notification status updated to read"
          },
          "404": {
            "description": "Notification not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/unread/{notificationId}": {
      "put": {
        "tags": [
          "Notifications"
        ],
        "summary": "Update notification to unread",
        "description": "Update notification to unread",
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Notification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Notification status updated to unread"
          },
          "404": {
            "description": "Notification not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  }
}